新增資料夾/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
PyJWT==2.8.0
bcrypt==4.1.2
google-auth==2.23.4
requests==2.31.0

